+++ pages/transaction/[transaction].tsx
@@ -24,7 +24,7 @@ export default function Transaction() {
   return (
     <Layout>
       <Grid container className={classes.root}>
-        <Grid item xs={12} sm={8}>
+        <Grid item xs={12} md={7}>
           <TransactionDetails hashValue={router.query.transaction} />
         </Grid>
       </Grid>