+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-## Unreleased
+## [Unreleased]
+
+* Updated Grid size on medium screens to 7
+* Disabled back button on Page 1 in TablePagination
+* Updated the address of a miner in LatestBlocks
+
+## v0.1.0
 
 * Implemented page Layout 
 * Implemented Drawer, Search Bar, Footer, Price Card, Chart Data, Network Dropdown
@@ -31,4 +37,5 @@
 * Removed Block Confirmation value in Transaction Details 
 * [#43] Updated Pagination for LatestBlocks, LatestTransactions, AccountList, Downtime, Transactions, ValidatedBlocks, DepositList, ProposalVotingList, TokenHolders & ValidatorVotesList
 * [#21] Updated ChartData component height and fixed the layout to keep equal spacing between components 
-
+* [#30] Stored the GraphQL & account endpoint in environment variable
+* [#41] Updated Middle Ellipsis screen size function to useMediaQuery